+++ src/pythonwrench/re.py
@@ -9,7 +9,7 @@ from typing import Any, Callable, Iterable, List, Optional, TypeVar, Union
 
 from typing_extensions import TypeAlias
 
-from pythonwrench.collections import find
+from pythonwrench.collections import filter_iterable, find
 
 T = TypeVar("T")
 
@@ -98,3 +98,21 @@ def sort_with_patterns(
     key_fn = get_key_fn(patterns, match_fn=match_fn)
     x = sorted(x, key=key_fn, reverse=reverse)
     return x
+
+
+def filter_with_patterns(
+    x: Iterable[str],
+    include: Optional[PatternListLike] = ".*",
+    *,
+    exclude: Optional[PatternListLike] = (),
+    match_fn: Callable[[PatternLike, str], Any] = re.search,
+) -> List[str]:
+    if include is None:
+        include = ".*"
+    if exclude is None:
+        exclude = ()
+    return [
+        xi
+        for xi in x
+        if match_patterns(xi, include, exclude=exclude, match_fn=match_fn)
+    ]