+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 - `as_builtin` now supports `datetime.date` instances.
 - `check_only_first` argument in `isinstance_generic` function.
 - `filter_with_patterns` function.
-- `skipfiles` and `include` for `tree_iter` function.
+- `skipfiles`, `include` and `sort` to `tree_iter` function.
 - `SupportsGetitem` protocols now has `T_Index` typevar set to `Any` instead of `int` by default.
 
 ## [0.4.2] 2025-10-16