+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.4.3] UNRELEASED
+## [0.4.3] 2025-12-13
 ### Added
 - `as_builtin` now supports `datetime.date` instances.
 - `check_only_first` argument in `isinstance_generic` function.